+++ d2/src/main/java/com/linkedin/d2/balancer/strategies/relative/RelativeLoadBalancerStrategy.java
@@ -96,6 +96,11 @@ public class RelativeLoadBalancerStrategy implements LoadBalancerStrategy
     return _stateUpdater.getPartitionState(partitionId);
   }
 
+  public int getFirstValidPartitionId()
+  {
+    return _stateUpdater.getFirstValidPartitionId();
+  }
+
   /**
    * Exposed for testings
    */