+++ CHANGELOG.md
@@ -13,9 +13,24 @@ When updating the changelog, remember to be very clear about what behavior has c
 and what APIs have changed, if applicable.
 
 ## [Unreleased]
+
+## [29.7.11] - 2020-10-19
+- Clear the destination directory for generateRestClientTask before the task runs.
+- Add 'ExtensionSchemaAnnotationHandler' for extension schema annotation compatibility check
+- Set javac source and target compatibility of dataTemplate compile task to "1.8" as the cogen changes in 29.7.0 is using Java 8 features.
+
+## [29.7.10] - 2020-10-15
+- Minimize memory copies and object creation during encoding.
+- Use String switch instead of map lookup in traverse callback for better performance
+- Reset isTraversing when cloning
+- Cache data objects in wrapped mapped/lists lazily on get.
+- Compute dataComplexHashCode lazily for DataList and DataMap
+
+## [29.7.9] - 2020-10-15
+- Add partition validation when getting relative load balancer metrics.
 - Extend checkPegasusSchemaSnapshot task to be enable to check schema annotation compatibility.
-  - The annotation compatibility will be triggered if SchemaANnotationHandler config is provided.
-  - Update SchemaAnnotationHandler interface to have a new api - annotationCompatibilityCheck, which can be used to check the custom annotation compatibility check.
+- The annotation compatibility will be triggered if SchemaANnotationHandler config is provided.
+- Update SchemaAnnotationHandler interface to have a new api - annotationCompatibilityCheck, which can be used to check the custom annotation compatibility check.
 
 ## [29.7.8] - 2020-10-12
 - Encoding performance improvements
@@ -109,7 +124,6 @@ and what APIs have changed, if applicable.
 ## [29.5.5] - 2020-08-21
 - Updated File and class path DataSchemaResolvers to resolve extension schemas from `/extensions` directory if specified.
 - Added `DarkGateKeeper` to enable users to provide custom implementation to determine if requests are to be dispatched to dark clusters.
->>>>>>> upstream/master
 
 ## [29.5.4] - 2020-08-17
 - Increase default timeout for symbol table fetch to 1s.
@@ -4700,7 +4714,10 @@ patch operations can re-use these classes for generating patch messages.
 
 ## [0.14.1]
 
-[Unreleased]: https://github.com/linkedin/rest.li/compare/v29.7.8...master
+[Unreleased]: https://github.com/linkedin/rest.li/compare/v29.7.11...master
+[29.7.11]: https://github.com/linkedin/rest.li/compare/v29.7.10...v29.7.11
+[29.7.10]: https://github.com/linkedin/rest.li/compare/v29.7.9...v29.7.10
+[29.7.9]: https://github.com/linkedin/rest.li/compare/v29.7.8...v29.7.9
 [29.7.8]: https://github.com/linkedin/rest.li/compare/v29.7.7...v29.7.8
 [29.7.7]: https://github.com/linkedin/rest.li/compare/v29.7.6...v29.7.7
 [29.7.6]: https://github.com/linkedin/rest.li/compare/v29.7.5...v29.7.6